주석 수정 및 페이징 리팩토링

+++ src/main/java/com/schedulerprojectdevelop/domain/comment/repository/CommentRepository.java
@@ -12,5 +12,5 @@ public interface CommentRepository extends JpaRepository<Comment, Long> {
 
     void deleteByUser_Id(Long userId);
 
-    int countBySchedule_Id(Long scheduleId);
+    Long countBySchedule_Id(Long scheduleId);
 }
